Move most of the query DSL implementation (predicates) to the engine
