+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/search/impl/SearchQueryResultDefinitionContextImpl.java
@@ -20,7 +20,7 @@ import org.hibernate.search.v6poc.backend.elasticsearch.document.model.impl.Elas
 import org.hibernate.search.v6poc.backend.elasticsearch.document.model.impl.ElasticsearchIndexModel;
 import org.hibernate.search.v6poc.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl.ElasticsearchSearchTargetContext;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl.SearchQueryWrappingDefinitionResultContextImpl;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchQueryWrappingDefinitionResultContextImpl;
 import org.hibernate.search.v6poc.engine.spi.SessionContext;
 import org.hibernate.search.v6poc.search.DocumentReference;
 import org.hibernate.search.v6poc.search.ObjectLoader;
