+++ engine/src/main/java/org/hibernate/search/v6poc/search/dsl/query/impl/SearchQueryContextImpl.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl;
+package org.hibernate.search.v6poc.search.dsl.query.impl;
 
 import java.util.Collection;
 import java.util.function.Function;
 
-import org.hibernate.search.v6poc.backend.elasticsearch.search.impl.SearchQueryBuilder;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchQueryBuilder;
 import org.hibernate.search.v6poc.search.SearchQuery;
 import org.hibernate.search.v6poc.search.dsl.query.SearchQueryContext;
 
@@ -17,7 +17,7 @@ import org.hibernate.search.v6poc.search.dsl.query.SearchQueryContext;
 /**
  * @author Yoann Rodiere
  */
-class SearchQueryContextImpl<T, Q, C> implements SearchQueryContext<Q> {
+public class SearchQueryContextImpl<T, Q, C> implements SearchQueryContext<Q> {
 
 	private final SearchQueryBuilder<T, C> searchQueryBuilder;
